+++ packages/iTwinUI-react/src/core/index.ts
@@ -59,6 +59,15 @@ export type { ComboBoxProps } from './ComboBox';
 export { DatePicker, generateLocalizedStrings } from './DatePicker';
 export type { DatePickerProps } from './DatePicker';
 
+export { Dialog } from './Dialog';
+export type {
+  DialogProps,
+  DialogButtonBarProps,
+  DialogContentProps,
+  DialogMainProps,
+  DialogTitleBarProps,
+} from './Dialog';
+
 export { DropdownMenu } from './DropdownMenu';
 export type { DropdownMenuProps } from './DropdownMenu';
 
@@ -165,7 +174,12 @@ export { RadioTile, RadioTileGroup } from './RadioTiles';
 export type { RadioTileGroupProps, RadioTileProps } from './RadioTiles';
 
 export { Select } from './Select';
-export type { SelectProps, SelectOption, ItemRendererProps } from './Select';
+export type {
+  SelectProps,
+  SelectOption,
+  ItemRendererProps,
+  SelectValueChangeEvent,
+} from './Select';
 
 export {
   SideNavigation,